requirements.txt
bcrypt==4.1.2
cffi==1.16.0
cryptography==42.0.3
iniconfig==2.0.0
limelightlib-python==0.9.1
packaging==23.2
navx==3.4.0
phoenix6==24.2.0
photonlibpy==2024.2.6
Pint==0.23
pluggy==1.4.0
pycparser==2.21
pyfrc==2024.0.1
PyNaCl==1.5.0
pynetconsole==2.0.4
pyntcore==2024.2.1.2
pytest==8.0.0
pytest-reraise==2.1.2
robotpy==2024.2.1.2
robotpy-apriltag==2024.2.1.2
robotpy-cli==2024.0.0
robotpy-commands-v2==2024.2.2
robotpy-cscore==2024.2.1.2
robotpy-ctre==2024.1.2
robotpy-hal==2024.2.1.2
robotpy-halsim-ds-socket==2024.2.1.2
robotpy-halsim-gui==2024.2.1.2
robotpy-halsim-ws==2024.2.1.2
robotpy-installer==2024.2.0
robotpy-navx==2024.1.0
robotpy-pathplannerlib==2024.2.2
robotpy-playingwithfusion==2024.1.0
robotpy-rev==2024.2.1
robotpy-romi==2024.2.1.2
robotpy-wpilib-utilities==2024.0.0
robotpy-wpimath==2024.2.1.2
robotpy-wpinet==2024.2.1.2
robotpy-wpiutil==2024.2.1.2
robotpy-xrp==2024.2.1.2
tomli==2.0.1
typing_extensions==4.9.0
websocket-client==1.7.0
wpilib==2024.2.1.2
